+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 Changes:
 
+* feature: validate value of `protocol` tag on a Dataplane resource
+  [#576](https://github.com/Kong/kuma/pull/576)
 * feature: support `<port>.service.kuma.io/protocol` annotation on k8s as a way for users to indicate protocol of a service
   [#575](https://github.com/Kong/kuma/pull/575)
 * feature: generate HTTP-specific inbound listeners for services tagged with `protocol: http`